open_data/pyproject.toml
[tool.poetry]
name = "opendata"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
requests = "^2.25.1"
Pillow = "^8.2.0"
dictor = "^0.1.7"
django-colorfield = "^0.4.1"
django-nested-inline = "^0.4.2"
django-notifications-hq = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
